// src/app/models/branch.model.ts
export class branchmodel {
  SelectedField: string;
  SelectedOperator: string;
  SelectedInput: string;
  selectedSaprator: string;
  selectedAndOr: string;
  SelectedAndOr: string;
  expression: string;
  SelectedColumnName: string;
  datatypevalue: string;
  Techexpression: string;


  ResultantExpression: string;
  ResultantTechnicalExpression: string;

}